Crate simple_model

Re-exports

pub use substance::Substance;

pub use hvac::HVAC;

Modules

boundary

building

construction

fenestration

hvac

infiltration

luminaire

material

model

rhai_api

scanner

simulation_state

simulation_state_element

space

substance

surface

Structs

Building

This object is utilized to group Space objects together for metering and/or shared values. For example, the number of storeys and the ShelterClass will help defining the Infiltrations

Construction

An object representing a multilayer Construction; that is to say, an array of Materials

Fenestration

A surface that can potentially be opened and closed. It can be of any Construction and it does not need to be a hole in another surface.

Luminaire

A Luminaire

Material

The representation of a physical layer-Material. That is to say, a layer of a certain thickness made of a certain Substance

SimpleModel

SimulationStateHeader

The SimulationState is a Vector of SimulationStateElement objects. It is intended to be a quick-to-clone structure.

Space

Represents a space within a building. This will often be a room, but it might also be half a room

Surface

A fixed surface in the building (or surroundings). This can be of any Construction, transparent or not.

Enums

Boundary

Represents the boundary of a Surface

FenestrationPositions

FenestrationType

Infiltration

An infiltration rate for a Space

ShelterClass

SimulationStateElement

The idea is to have a cheap-to-clone (or copy?) structure

Type Definitions

Float

SimulationState