Struct simple_model::building::Building

pub struct Building {
    pub name: String,
    index: Option<usize>,
    n_storeys: Option<usize>,
    shelter_class: Option<ShelterClass>,
    stack_coefficient: Option<f64>,
    wind_coefficient: Option<f64>,
}

This object is utilized to group Space objects together for metering and/or shared values. For example, the number of storeys and the ShelterClass will help defining the Infiltrations

Fields
name: String

The name of the Building

index: Option<usize>

The index of the building in the SimpleModel’s buildings array

n_storeys: Option<usize>

The number of storeys of this building.

This value use used by the AirFlow module when a Space associated to this Building has been assigned an EffectiveAirLeakageArea infiltration. This value is required for calculating the Stack Coefficient ($C_s$) and the Wind Coefficient ($C_w$) of the EffectiveAirLeakageArea infiltration. $C_s$ and $C_w$ can be inputed directly by assigning values to the stack_coefficient and wind_coefficient fields, in which case the n_storeys field will be ignored.

shelter_class: Option<ShelterClass>

The ShelterClass of this building.

This value use used by the AirFlow module when a Space associated to this Building has been assigned an EffectiveAirLeakageArea infiltration. This value is required for calculating the Wind Coefficient ($C_s$) of the EffectiveAirLeakageArea infiltration. $C_w$ can be inputed directly by assigning values to the wind_coefficient field, in which case the shelter_class field will be ignored.

stack_coefficient: Option<f64>

The stack coefficient of this building, used for calculating infiltrations in Spaces that utilize the EffectiveAirLeakageArea infiltration option.

If not given, the number of storeys will be used for getting this values (based on EnergyPlus’ Engineering Reference).

Note: The EffectiveAirLeakageArea object is appropriate for buildings of 3 storeys or less.

wind_coefficient: Option<f64>

The wind coefficient of this building, used for calculating infiltrations in Spaces that utilize the EffectiveAirLeakageArea infiltration option.

If not given, the number of storeys will be used for getting this values (based on EnergyPlus’ Engineering Reference).

Note: The EffectiveAirLeakageArea object is appropriate for buildings of 3 storeys or less.

Implementations

impl Building

pub fn from_bytes(
line: usize,
bytes: &[u8],
model: &SimpleModel
) -> Result<Self, String>

Creates a new Building by scanning bytes (i.e., a &[u8]) acquired from a text file

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new(name: String) -> Self

Creates a new Building

All the required fields are asked by the constructor. The Optional fields default to None

pub fn index(&self) -> Result<&usize, String>

Gets the index field. Returns a Result because this field is optional and thus it might not be there.

pub fn n_storeys(&self) -> Result<&usize, String>

Gets the n_storeys field. Returns a Result because this field is optional and thus it might not be there.

pub fn shelter_class(&self) -> Result<&ShelterClass, String>

Gets the shelter_class field. Returns a Result because this field is optional and thus it might not be there.

pub fn stack_coefficient(&self) -> Result<&f64, String>

Gets the stack_coefficient field. Returns a Result because this field is optional and thus it might not be there.

pub fn wind_coefficient(&self) -> Result<&f64, String>

Gets the wind_coefficient field. Returns a Result because this field is optional and thus it might not be there.

pub fn set_index(&mut self, v: usize) -> &mut Self

Sets the index field.

pub fn set_n_storeys(&mut self, v: usize) -> &mut Self

Sets the n_storeys field.

pub fn set_shelter_class(&mut self, v: ShelterClass) -> &mut Self

Sets the shelter_class field.

pub fn set_stack_coefficient(&mut self, v: f64) -> &mut Self

Sets the stack_coefficient field.

pub fn set_wind_coefficient(&mut self, v: f64) -> &mut Self

Sets the wind_coefficient field.