Enum model::Output

pub enum Output {

    Clothing(String),
    FenestrationOpenFraction(String),
    HeatingCoolingPowerConsumption(String),
    LuminairePowerConsumption(String),
    SurfaceFrontConvectionCoefficient(String),
    SurfaceBackConvectionCoefficient(String),
    SurfaceFrontConvectiveHeatFlow(String),
    SurfaceBackConvectiveHeatFlow(String),
    SurfaceFrontSolarIrradiance(String),
    SurfaceBackSolarIrradiance(String),
    SurfaceFrontIRIrradiance(String),
    SurfaceBackIRIrradiance(String),
    FenestrationFrontConvectionCoefficient(String),
    FenestrationBackConvectionCoefficient(String),
    FenestrationFrontConvectiveHeatFlow(String),
    FenestrationBackConvectiveHeatFlow(String),
    FenestrationFrontSolarIrradiance(String),
    FenestrationBackSolarIrradiance(String),
    FenestrationFrontIRIrradiance(String),
    FenestrationBackIRIrradiance(String),
    SpaceDryBulbTemperature(String),
    SpaceInfiltrationVolume(String),
    SpaceInfiltrationTemperature(String),
    SpaceVentilationVolume(String),
    SpaceVentilationTemperature(String),
    SpaceAirExchangeVolume(String),
    SurfaceNodeTemperature(String),
    FenestrationNodeTemperature(String),

}

Possible outputs to request from the simulation

Example

{{#include ../../../tests/box.spl:bedroom}}
{{#include ../../../tests/box.spl:bedroom_output}}

Variants

Clothing(String)

The amount of clothing the person is using, in Clo value

FenestrationOpenFraction(String)

Represents how open is a fenestration. Contains the Index of fenestration, and its open fraction

HeatingCoolingPowerConsumption(String)

Represents the heating/cooling energy consumption of a Heating/Cooling system, in Watts

Contains the index of the HVAC in the building's vector, and the power.

LuminairePowerConsumption(String)

Represents the power being consumed by a Luminaire object, in Watts (luminaire index, power)

SurfaceFrontConvectionCoefficient(String)

The convective heat transfer coefficient at the front of a surface

SurfaceBackConvectionCoefficient(String)

The convective heat transfer coefficient at the back of a surface

SurfaceFrontConvectiveHeatFlow(String)

The convective heat flow at the front of a surface

SurfaceBackConvectiveHeatFlow(String)

The convective heat flow at the back of a surface

SurfaceFrontSolarIrradiance(String)

Incident solar irradiance at the front

SurfaceBackSolarIrradiance(String)

Incident solar irradiance at the back

SurfaceFrontIRIrradiance(String)

Incident Infrared irradiance at the front

SurfaceBackIRIrradiance(String)

Incident Infrared irradiance at the back

FenestrationFrontConvectionCoefficient(String)

The convective heat transfer coefficient at the front of a surface

FenestrationBackConvectionCoefficient(String)

The convective heat transfer coefficient at the back of a surface

FenestrationFrontConvectiveHeatFlow(String)

The convective heat flow at the front of a surface

FenestrationBackConvectiveHeatFlow(String)

The convective heat flow at the back of a surface

FenestrationFrontSolarIrradiance(String)

Incident solar irradiance at the front

FenestrationBackSolarIrradiance(String)

Incident solar irradiance at the back

FenestrationFrontIRIrradiance(String)

Incident Infrared irradiance at the front

FenestrationBackIRIrradiance(String)

Incident Infrared irradiance at the back

SpaceDryBulbTemperature(String)

Space Air Temperature in C… The elements are the index of the Space in the Building mode and the temperature

SpaceInfiltrationVolume(String)

The volume of air that is entering the space in an uncontrolled way. In m3/s

SpaceInfiltrationTemperature(String)

The temperature of air that is entering the space in an uncontrolled way. In C

SpaceVentilationVolume(String)

The volume of air that is entering the space in a controlled way. In m3/s

SpaceVentilationTemperature(String)

The temperature of air that is entering the space in a controlled way. In C

SpaceAirExchangeVolume(String)

The volume of air that is moving from one space to another in a controlled way. In m3/s

SurfaceNodeTemperature(String)

Temperature (Float) of Surface's (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

FenestrationNodeTemperature(String)

Temperature (Float) of Fenestration's (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

Implementations

impl Output

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

Prints the docs for this object

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects