Enum simple_model::SimulationStateElement

pub enum SimulationStateElement {

    Clothing,
    FenestrationOpenFraction(usize),
    HeatingCoolingPowerConsumption(usize),
    LuminairePowerConsumption(usize),
    SurfaceFrontConvectionCoefficient(usize),
    SurfaceBackConvectionCoefficient(usize),
    SurfaceFrontConvectiveHeatFlow(usize),
    SurfaceBackConvectiveHeatFlow(usize),
    SurfaceFrontSolarIrradiance(usize),
    SurfaceBackSolarIrradiance(usize),
    SurfaceFrontIRIrradiance(usize),
    SurfaceBackIRIrradiance(usize),
    FenestrationFrontConvectionCoefficient(usize),
    FenestrationBackConvectionCoefficient(usize),
    FenestrationFrontConvectiveHeatFlow(usize),
    FenestrationBackConvectiveHeatFlow(usize),
    FenestrationFrontSolarIrradiance(usize),
    FenestrationBackSolarIrradiance(usize),
    FenestrationFrontIRIrradiance(usize),
    FenestrationBackIRIrradiance(usize),
    SpaceDryBulbTemperature(usize),
    SpaceInfiltrationVolume(usize),
    SpaceInfiltrationTemperature(usize),
    SpaceVentilationVolume(usize),
    SpaceVentilationTemperature(usize),
    SpaceAirExchangeVolume(usize, usize),
    SurfaceNodeTemperature(usize, usize),
    FenestrationNodeTemperature(usize, usize),

}

The idea is to have a cheap-to-clone (or copy?) structure

Variants

Clothing

The amount of clothing the person is using, in Clo value

FenestrationOpenFraction(usize)

Tuple Fields

0: usize

Represents how open is a fenestration. Contains the Index of fenestration, and its open fraction

HeatingCoolingPowerConsumption(usize)

Tuple Fields

0: usize

Represents the heating/cooling energy consumption of a Heating/Cooling system, in Watts

Contains the index of the HVAC in the building’s vector, and the power.

LuminairePowerConsumption(usize)

Tuple Fields

0: usize

Represents the power being consumed by a Luminaire object, in Watts (luminaire index, power)

SurfaceFrontConvectionCoefficient(usize)

Tuple Fields

0: usize

The convective heat transfer coefficient at the front of a surface

SurfaceBackConvectionCoefficient(usize)

Tuple Fields

0: usize

The convective heat transfer coefficient at the back of a surface

SurfaceFrontConvectiveHeatFlow(usize)

Tuple Fields

0: usize

The convective heat flow at the front of a surface

SurfaceBackConvectiveHeatFlow(usize)

Tuple Fields

0: usize

The convective heat flow at the back of a surface

SurfaceFrontSolarIrradiance(usize)

Tuple Fields

0: usize

Incident solar irradiance at the front

SurfaceBackSolarIrradiance(usize)

Tuple Fields

0: usize

Incident solar irradiance at the back

SurfaceFrontIRIrradiance(usize)

Tuple Fields

0: usize

Incident Infrared irradiance at the front

SurfaceBackIRIrradiance(usize)

Tuple Fields

0: usize

Incident Infrared irradiance at the back

FenestrationFrontConvectionCoefficient(usize)

Tuple Fields

0: usize

The convective heat transfer coefficient at the front of a surface

FenestrationBackConvectionCoefficient(usize)

Tuple Fields

0: usize

The convective heat transfer coefficient at the back of a surface

FenestrationFrontConvectiveHeatFlow(usize)

Tuple Fields

0: usize

The convective heat flow at the front of a surface

FenestrationBackConvectiveHeatFlow(usize)

Tuple Fields

0: usize

The convective heat flow at the back of a surface

FenestrationFrontSolarIrradiance(usize)

Tuple Fields

0: usize

Incident solar irradiance at the front

FenestrationBackSolarIrradiance(usize)

Tuple Fields

0: usize

Incident solar irradiance at the back

FenestrationFrontIRIrradiance(usize)

Tuple Fields

0: usize

Incident Infrared irradiance at the front

FenestrationBackIRIrradiance(usize)

Tuple Fields

0: usize

Incident Infrared irradiance at the back

SpaceDryBulbTemperature(usize)

Tuple Fields

0: usize

Space Air Temperature in C… The elements are the index of the Space in the Building mode and the temperature

SpaceInfiltrationVolume(usize)

Tuple Fields

0: usize

The volume of air that is entering the space in an uncontrolled way. In m3/s

SpaceInfiltrationTemperature(usize)

Tuple Fields

0: usize

The temperature of air that is entering the space in an uncontrolled way. In C

SpaceVentilationVolume(usize)

Tuple Fields

0: usize

The volume of air that is entering the space in a controlled way. In m3/s

SpaceVentilationTemperature(usize)

Tuple Fields

0: usize

The temperature of air that is entering the space in a controlled way. In C

SpaceAirExchangeVolume(usize, usize)

Tuple Fields

0: usize

1: usize

The volume of air that is moving from one space to another in a controlled way. In m3/s

SurfaceNodeTemperature(usize, usize)

Tuple Fields

0: usize

1: usize

Temperature (Float) of Surface’s (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

FenestrationNodeTemperature(usize, usize)

Tuple Fields

0: usize

1: usize

Temperature (Float) of Fenestration’s (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

Implementations

impl SimulationStateElement

pub fn is_physical(&self) -> bool

Checks whether a SimulationStateElement is of kind Physical

pub fn is_personal(&self) -> bool

Checks whether a SimulationStateElement is of kind Personal

pub fn is_operational(&self) -> bool

Checks whether a SimulationStateElement is of kind Operational