Struct simple_model::fenestration::Fenestration

pub struct Fenestration {

    pub name: String,
    index: Option<usize>,
    pub vertices: Polygon3D,
    pub construction: Rc<Construction>,
    pub operation_type: FenestrationPositions,
    pub fenestration_type: FenestrationType,
    front_boundary: Option<Boundary>,
    back_boundary: Option<Boundary>,
    first_node_temperature: RefCell<Option<usize>>,
    last_node_temperature: RefCell<Option<usize>>,
    open_fraction: RefCell<Option<usize>>,
    front_convection_coefficient: RefCell<Option<usize>>,
    back_convection_coefficient: RefCell<Option<usize>>,
    front_convective_heat_flow: RefCell<Option<usize>>,
    back_convective_heat_flow: RefCell<Option<usize>>,
    front_incident_solar_irradiance: RefCell<Option<usize>>,
    back_incident_solar_irradiance: RefCell<Option<usize>>,
    front_ir_irradiance: RefCell<Option<usize>>,
    back_ir_irradiance: RefCell<Option<usize>>,

}

A surface that can potentially be opened and closed. It can be of any Construction and it does not need to be a hole in another surface.

Fields
name: String

The name of the sub surface

index: Option<usize>

The position of this object in its contaner Vector

vertices: Polygon3D

An array of Numbers representing the vertices of the surface. The length of this array must be divisible by 3.

construction: Rc<Construction>

The index of the Construction object in the constructions property of the SimpleModel object

operation_type: FenestrationPositions

The opportunity for operating the Fenestration

fenestration_type: FenestrationType

It it a window or a door, or…?

front_boundary: Option<Boundary>

A reference to the Boundary in front of the Fenestration

back_boundary: Option<Boundary>

A reference to the Boundary in back of the Fenestration

first_node_temperature: RefCell<Option<usize>>last_node_temperature: RefCell<Option<usize>>open_fraction: RefCell<Option<usize>>

Index of the SimulationStateElement representing the fraction open in the SimulationState

front_convection_coefficient: RefCell<Option<usize>>back_convection_coefficient: RefCell<Option<usize>>front_convective_heat_flow: RefCell<Option<usize>>back_convective_heat_flow: RefCell<Option<usize>>front_incident_solar_irradiance: RefCell<Option<usize>>back_incident_solar_irradiance: RefCell<Option<usize>>front_ir_irradiance: RefCell<Option<usize>>back_ir_irradiance: RefCell<Option<usize>>
Implementations

impl Fenestration

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new(
name: String,
vertices: Polygon3D,
construction: Rc<Construction>,
operation_type: FenestrationPositions,
fenestration_type: FenestrationType
) -> Self

Creates a new Fenestration

All the required fields are asked by the constructor. The Optional fields default to None

pub fn index(&self) -> Result<&usize, String>

Gets the index field. Returns a Result because this field is optional and thus it might not be there.

pub fn front_boundary(&self) -> Result<&Boundary, String>

Gets the front_boundary field. Returns a Result because this field is optional and thus it might not be there.

pub fn back_boundary(&self) -> Result<&Boundary, String>

Gets the back_boundary field. Returns a Result because this field is optional and thus it might not be there.

pub fn set_index(&mut self, v: usize) -> &mut Self

Sets the index field.

pub fn set_front_boundary(&mut self, v: Boundary) -> &mut Self

Sets the front_boundary field.

pub fn set_back_boundary(&mut self, v: Boundary) -> &mut Self

Sets the back_boundary field.

pub fn first_node_temperature_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the first_node_temperature within the SimulationState.

pub fn first_node_temperature(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the first_node_temperature within the SimulationState.

pub fn last_node_temperature_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the last_node_temperature within the SimulationState.

pub fn last_node_temperature(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the last_node_temperature within the SimulationState.

pub fn open_fraction_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the open_fraction within the SimulationState.

pub fn open_fraction(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the open_fraction within the SimulationState.

pub fn front_convection_coefficient_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the front_convection_coefficient within the SimulationState.

pub fn front_convection_coefficient(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the SimulationStateElement representing the front_convection_coefficient within the SimulationState.

pub fn back_convection_coefficient_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the back_convection_coefficient within the SimulationState.

pub fn back_convection_coefficient(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the SimulationStateElement representing the back_convection_coefficient within the SimulationState.

pub fn front_convective_heat_flow_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the front_convective_heat_flow within the SimulationState.

pub fn front_convective_heat_flow(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the front_convective_heat_flow within the SimulationState.

pub fn back_convective_heat_flow_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the back_convective_heat_flow within the SimulationState.

pub fn back_convective_heat_flow(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the back_convective_heat_flow within the SimulationState.

pub fn front_incident_solar_irradiance_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the front_incident_solar_irradiance within the SimulationState.

pub fn front_incident_solar_irradiance(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the SimulationStateElement representing the front_incident_solar_irradiance within the SimulationState.

pub fn back_incident_solar_irradiance_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the back_incident_solar_irradiance within the SimulationState.

pub fn back_incident_solar_irradiance(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the SimulationStateElement representing the back_incident_solar_irradiance within the SimulationState.

pub fn front_ir_irradiance_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the front_ir_irradiance within the SimulationState.

pub fn front_ir_irradiance(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the front_ir_irradiance within the SimulationState.

pub fn back_ir_irradiance_index(&self) -> Option<usize>

Gets the index of the SimulationStateElement representing the back_ir_irradiance within the SimulationState.

pub fn back_ir_irradiance(&self, state: &SimulationState) -> Option<f64>

Gets the value of the SimulationStateElement representing the back_ir_irradiance within the SimulationState.

pub fn set_first_node_temperature_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the first_node_temperature within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_first_node_temperature(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the first_node_temperature within the SimulationState .

pub fn set_last_node_temperature_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the last_node_temperature within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_last_node_temperature(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the last_node_temperature within the SimulationState .

pub fn set_open_fraction_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the open_fraction within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_open_fraction(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the open_fraction within the SimulationState .

pub fn set_front_convection_coefficient_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the front_convection_coefficient within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_front_convection_coefficient(
&self,
state: &mut SimulationState,
v: f64
)

Changes the value of the SimulationStateElement associated with the front_convection_coefficient within the SimulationState .

pub fn set_back_convection_coefficient_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the back_convection_coefficient within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_back_convection_coefficient(
&self,
state: &mut SimulationState,
v: f64
)

Changes the value of the SimulationStateElement associated with the back_convection_coefficient within the SimulationState .

pub fn set_front_convective_heat_flow_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the front_convective_heat_flow within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_front_convective_heat_flow(
&self,
state: &mut SimulationState,
v: f64
)

Changes the value of the SimulationStateElement associated with the front_convective_heat_flow within the SimulationState .

pub fn set_back_convective_heat_flow_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the back_convective_heat_flow within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_back_convective_heat_flow(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the back_convective_heat_flow within the SimulationState .

pub fn set_front_incident_solar_irradiance_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the front_incident_solar_irradiance within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_front_incident_solar_irradiance(
&self,
state: &mut SimulationState,
v: f64
)

Changes the value of the SimulationStateElement associated with the front_incident_solar_irradiance within the SimulationState .

pub fn set_back_incident_solar_irradiance_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the back_incident_solar_irradiance within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_back_incident_solar_irradiance(
&self,
state: &mut SimulationState,
v: f64
)

Changes the value of the SimulationStateElement associated with the back_incident_solar_irradiance within the SimulationState .

pub fn set_front_ir_irradiance_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the front_ir_irradiance within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_front_ir_irradiance(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the front_ir_irradiance within the SimulationState .

pub fn set_back_ir_irradiance_index(&self, i: usize)

Sets the index of the SimulationStateElement representing the back_ir_irradiance within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_back_ir_irradiance(&self, state: &mut SimulationState, v: f64)

Changes the value of the SimulationStateElement associated with the back_ir_irradiance within the SimulationState .

impl Fenestration

pub fn from_bytes(
line: usize,
bytes: &[u8],
model: &SimpleModel
) -> Result<Self, String>

Creates a new Fenestration by scanning bytes (i.e., a &[u8]) acquired from a text file

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

impl Fenestration

pub fn register_api(
engine: &mut Engine,
model: &Rc<SimpleModel>,
state: &Rc<RefCell<SimulationState>>,
research_mode: bool
)

pub fn print_api_doc(dir: &str, summary: &mut String) -> Result<()>

impl Fenestration

pub fn clone_loop(&self) -> Loop3D

Clones the outer [Loop3D] of the Fenestration

pub fn area(&self) -> f64

Gets the area, based on the [Polygon3D] that represents this Fenestration

pub fn is_operable(&self) -> bool