Struct simple_model::luminaire::Luminaire

pub struct Luminaire { /* fields omitted */ }

A Luminaire

Please fill this doc

Implementations

impl Luminaire

pub fn from_bytes(bytes: &[u8], model: &mut SimpleModel) -> Result<Self, String>

Creates a new Luminaire by scanning bytes (i.e., a &[u8]) acquired from a text file

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

impl Luminaire

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn new(name: String) -> Self

Creates a new Luminaire

All the required fields are asked by the constructor. The Optional fields default to None

pub fn index(&self) -> Result<&usize, String>

Gets the index field. Returns a Result because this field is optional and thus it might not be there.

pub fn max_power(&self) -> Result<&f64, String>

Gets the max_power field. Returns a Result because this field is optional and thus it might not be there.

pub fn target_space(&self) -> Result<&Rc<Space>, String>

Gets the target_space field. Returns a Result because this field is optional and thus it might not be there.

pub fn set_index(&mut self, v: usize) -> &mut Self

Sets the index field.

pub fn set_max_power(&mut self, v: f64) -> &mut Self

Sets the max_power field.

pub fn set_target_space(&mut self, v: Rc<Space>) -> &mut Self

Sets the target_space field.

pub fn power_consumption_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the power_consumption within the SimulationState.

pub fn power_consumption(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the power_consumption within the SimulationState.

pub fn set_power_consumption_index(&self, i: usize)

Sets the index of the [SimulationStateElement] representing the power_consumption within the SimulationState. Will panic if the object already has an index assigned for this field.

pub fn set_power_consumption(&self, state: &mut SimulationState, v: f64)

Changes the value of the [SimulationStateElement] associated with the power_consumption within the SimulationState .