Struct simple_model::scanner::simple_scanner::SimpleScanner

pub struct SimpleScanner<'a> {
    pub line: usize,
    source: &'a [u8],
    current_index: usize,
    start_index: usize,
    pub error_msg: String,
    finished: bool,
}

Fields
line: usizesource: &'a [u8]current_index: usizestart_index: usizeerror_msg: Stringfinished: bool
Implementations

impl<'a> SimpleScanner<'a>

pub fn new(source: &'a [u8], line: usize) -> Self

Creates a new SimpleScanner

pub fn borrow_slice(&self, start: usize, end: usize) -> &[u8]

pub fn update_start_index(&mut self)

fn make_token(&self, token_type: TokenType) -> Token<'a>

Builds a Token corresponding to the current position of the scanner (e.g., start_index and current_index, line, etc.)

fn match_char(&mut self, expected: char) -> bool

Checks whether the char at the current_index in the source is equal to the expected char. If it is, it returns true and advances one step. If not, it returns false and does not advance.

fn advance(&mut self) -> Option<char>

Advances one char in the source, returning the consumed char inside of an Option. If finished, it will mark the SimpleScanner as finished and return None

pub fn peek(&self) -> char

Gets the char at the current_index. Returns \0 if finished.

fn peek_next(&self) -> char

Gets the char following the current_index. Returns \0 if finished.

fn number(&mut self) -> Token<'a>

Consumes a number, producing a Token

fn string(&mut self, delimiter: char) -> Token<'a>

Consumes a string (i.e., a term in “quotes”)

fn identifier(&mut self) -> Token<'a>

Consumes a whole Identifier

fn check_keyword(&self, word: &str) -> bool

Checks whether the chars between the SimpleScanners start_index and current_index
match a certain str (i.e., a keyword).

pub fn skip_white_space(&mut self)

Skips the white spaces and the comments and all those things.

pub fn scan_token(&mut self) -> Token<'a>

Scans a Token in SimpleBuilding read mode

pub fn get_object_slice(&mut self) -> (usize, usize)

Consumes an object and returns the start and end of that object.

pub fn scan_field(&mut self) -> Result<(Token<'_>, Token<'_>), String>

pub fn get_field(&mut self) -> Result<(Token<'_>, Token<'_>), String>

Checks whether a certain field is in the source code.

Note that this function is not meant to be fast, as it intends to search for a field within a small source.

pub fn parse_model(
&mut self
) -> Result<(SimpleModel, SimulationStateHeader), String>