Enum simple_model::simulation_state_element::SimulationStateElement

pub enum SimulationStateElement {

    Clothing,
    FenestrationOpenFraction(usize),
    HeatingCoolingPowerConsumption(usize),
    LuminairePowerConsumption(usize),
    SurfaceFrontConvectionCoefficient(usize),
    SurfaceBackConvectionCoefficient(usize),
    SurfaceFrontConvectiveHeatFlow(usize),
    SurfaceBackConvectiveHeatFlow(usize),
    SpaceDryBulbTemperature(usize),
    SpaceInfiltrationVolume(usize),
    SpaceInfiltrationTemperature(usize),
    SpaceVentilationVolume(usize),
    SpaceVentilationTemperature(usize),
    SpaceAirExchangeVolume(usize, usize),
    SurfaceNodeTemperature(usize, usize),
    FenestrationNodeTemperature(usize, usize),

}

The idea is to have a cheap-to-clone (or copy?) structure

Variants

Clothing

The amount of clothing the person is using, in Clo value

FenestrationOpenFraction(usize)

Represents how open is a fenestration. Contains the Index of fenestration, and its open fraction

Tuple Fields of FenestrationOpenFraction

0: usize

HeatingCoolingPowerConsumption(usize)

Represents the heating/cooling energy consumption of a Heating/Cooling system, in Watts

Contains the index of the HVAC in the building’s vector, and the power.

Tuple Fields of HeatingCoolingPowerConsumption

0: usize

LuminairePowerConsumption(usize)

Represents the power being consumed by a Luminaire object, in Watts (luminaire index, power)

Tuple Fields of LuminairePowerConsumption

0: usize

SurfaceFrontConvectionCoefficient(usize)

The convective heat transfer coefficient at the front of a surface

Tuple Fields of SurfaceFrontConvectionCoefficient

0: usize

SurfaceBackConvectionCoefficient(usize)

The convective heat transfer coefficient at the back of a surface

Tuple Fields of SurfaceBackConvectionCoefficient

0: usize

SurfaceFrontConvectiveHeatFlow(usize)

The convective heat flow at the front of a surface

Tuple Fields of SurfaceFrontConvectiveHeatFlow

0: usize

SurfaceBackConvectiveHeatFlow(usize)

The convective heat flow at the back of a surface

Tuple Fields of SurfaceBackConvectiveHeatFlow

0: usize

SpaceDryBulbTemperature(usize)

Space Air Temperature in C… The elements are the index of the Space in the Building mode and the temperature

Tuple Fields of SpaceDryBulbTemperature

0: usize

SpaceInfiltrationVolume(usize)

The volume of air that is entering the space in an uncontrolled way. In m3/s

Tuple Fields of SpaceInfiltrationVolume

0: usize

SpaceInfiltrationTemperature(usize)

The temperature of air that is entering the space in an uncontrolled way. In C

Tuple Fields of SpaceInfiltrationTemperature

0: usize

SpaceVentilationVolume(usize)

The volume of air that is entering the space in a controlled way. In m3/s

Tuple Fields of SpaceVentilationVolume

0: usize

SpaceVentilationTemperature(usize)

The temperature of air that is entering the space in a controlled way. In C

Tuple Fields of SpaceVentilationTemperature

0: usize

SpaceAirExchangeVolume(usize, usize)

The volume of air that is moving from one space to another in a controlled way. In m3/s

Tuple Fields of SpaceAirExchangeVolume

0: usize1: usize

SurfaceNodeTemperature(usize, usize)

Temperature (Float) of Surface’s (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

Tuple Fields of SurfaceNodeTemperature

0: usize1: usize

FenestrationNodeTemperature(usize, usize)

Temperature (Float) of Fenestration’s (usize) node (usize) I.e. the order is (Surface Index, Node index, Temperature).

Tuple Fields of FenestrationNodeTemperature

0: usize1: usize

Implementations

impl SimulationStateElement

pub fn is_physical(&self) -> bool

Checks whether a SimulationStateElement is of kind Physical

pub fn is_personal(&self) -> bool

Checks whether a SimulationStateElement is of kind Personal

pub fn is_operational(&self) -> bool

Checks whether a SimulationStateElement is of kind Operational