Struct simple_model::SimpleModel

pub struct SimpleModel {
    pub name: String,
    pub substances: Vec<Substance>,
    pub materials: Vec<Rc<Material>>,
    pub constructions: Vec<Rc<Construction>>,
    pub surfaces: Vec<Rc<Surface>>,
    pub spaces: Vec<Rc<Space>>,
    pub buildings: Vec<Rc<Building>>,
    pub fenestrations: Vec<Rc<Fenestration>>,
    pub hvacs: Vec<HVAC>,
    pub luminaires: Vec<Rc<Luminaire>>,
}

Fields
name: String

The name of the building

substances: Vec<Substance>materials: Vec<Rc<Material>>constructions: Vec<Rc<Construction>>surfaces: Vec<Rc<Surface>>spaces: Vec<Rc<Space>>buildings: Vec<Rc<Building>>fenestrations: Vec<Rc<Fenestration>>

The windows and doors in the surface

hvacs: Vec<HVAC>

The Heating/Cooling devices in the space

luminaires: Vec<Rc<Luminaire>>

Luminaires

Implementations

impl SimpleModel

pub fn from_bytes(
line: usize,
bytes: &[u8],
model: &SimpleModel
) -> Result<Self, String>

Creates a new SimpleModel by scanning bytes (i.e., a &[u8]) acquired from a text file

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new(name: String) -> Self

Creates a new SimpleModel

All the required fields are asked by the constructor. The Optional fields default to None

impl SimpleModel

pub fn from_file(
filename: String
) -> Result<(Self, SimulationStateHeader), String>

impl SimpleModel

pub fn add_building(&mut self, add: Building) -> Rc<Building>

Adds a Building to the SimpleModel

impl SimpleModel

pub fn add_construction(&mut self, add: Construction) -> Rc<Construction>

Adds a Construction to the SimpleModel

impl SimpleModel

pub fn add_material(&mut self, add: Material) -> Rc<Material>

Adds a Material to the SimpleModel

impl SimpleModel

pub fn add_substance(&mut self, add: Substance) -> Substance

Adds a Substance to the SimpleModel

impl SimpleModel

pub fn add_fenestration(
&mut self,
add: Fenestration,
state: &mut SimulationStateHeader
) -> Rc<Fenestration>

Adds a Fenestration to the SimpleModel

impl SimpleModel

pub fn add_luminaire(
&mut self,
add: Luminaire,
state: &mut SimulationStateHeader
) -> Rc<Luminaire>

Adds a Luminaire to the SimpleModel

impl SimpleModel

pub fn add_space(&mut self, add: Space) -> Rc<Space>

Adds a Space to the SimpleModel

impl SimpleModel

pub fn add_surface(&mut self, add: Surface) -> Rc<Surface>

Adds a Surface to the SimpleModel

impl SimpleModel

pub fn add_hvac(&mut self, add: HVAC, state: &mut SimulationStateHeader) -> HVAC

Adds a HVAC to the SimpleModel