Struct simple_model::SimulationStateHeader

pub struct SimulationStateHeader {
    n_operational: usize,
    n_individual: usize,
    passed_operational: bool,
    passed_personal: bool,
    pub elements: Vec<SimulationStateElement>,
    pub default_values: Option<SimulationState>,
}

The SimulationState is a Vector of SimulationStateElement objects. It is intended to be a quick-to-clone structure.

To make it quicker to operate, the following conventions are enforced:

Personal elements go first, then Operational go second, and Physical go third. (This is checked when pushing elements to the state)

Elements cannot be repeated (this is not really checked.).

Fields
n_operational: usize

The number of operational StateElements in the State

n_individual: usize

The number of individual StateElements in the State

passed_operational: bool

A flag indicating whether any physical element has been assigned. If it has, then no more individual elements are allowed. This allows maintaining (this is a convention, made in favour of performance) all the individual elements at the beginning, then all the operational ones, and then all the physical ones

passed_personal: bool

A flag indicating whether a non-individual element has been assigned. If it has, then no more individual elements are allowed. This allows maintaining (this is a convention, made in favour of performance) all the individual elements at the beginning, then all the operational ones, and then all the physical ones

elements: Vec<SimulationStateElement>

The elements in the SimulationState

default_values: Option<SimulationState>

The default values, from which the simulation will start

Implementations

impl SimulationStateHeader

pub fn new() -> Self

Creates a new empty Building State

pub fn push(&mut self, e: SimulationStateElement, value: f64) -> usize

Pushes an SimulationStateElement into the elements vector.

It ensures that operational elements are there first, and that physical are there afterwards.

Returns the index of the new element.

pub fn len(&self) -> usize

Returns the number of elements in the State element

pub fn is_empty(&self) -> bool

Checks if the State element is empty

pub fn n_operational(&self) -> usize

Returns the number of operational elements

pub fn n_individual(&self) -> usize

Returns the number of operational elements

pub fn copy_from(&mut self, origin: &SimulationStateHeader)

Copies the whole state into another state

pub fn copy_physical_state(
&mut self,
origin: &SimulationState,
destination: &mut SimulationState
)

Copies the Physical SimulationStateElements objects from origin to destination. It is expected that all the Operational BuildingStateElements are bundled at the beginning of the element.

pub fn copy_operational_state(
&mut self,
origin: &SimulationState,
destination: &mut SimulationState
)

Copies the Operational SimulationStateElements objects from origin to destination. It is expected that all the Operational BuildingStateElements are bundled at the beginning of the element.

pub fn copy_individual_state(
&mut self,
origin: &SimulationState,
destination: &mut SimulationState
)

Copies the Individual SimulationStateElements objects from origin to destination. It is expected that all the Operational BuildingStateElements are bundled at the beginning of the element.

pub fn take_values(&mut self) -> Option<SimulationState>

Takes the default_values from the SimulationStateHeader. Leaves None in its place.