Struct simple_model::substance::normal::Normal

pub struct Normal {
    pub name: String,
    index: Option<usize>,
    thermal_conductivity: Option<f64>,
    specific_heat_capacity: Option<f64>,
    density: Option<f64>,
}

Represents a Normal; that is to say, a physical materiality with physical properties. The name Normal has been chosen instead of Material to respect EnergyPlus’ and other software’s terminology (which does not include Substace, but it does include Material, which is essentially a Normal with a thickness).

Fields
name: String

The name of the Normal. Should be unique for each Material in the SimpleModel object

index: Option<usize>

The position of this object in its contaner Vector

thermal_conductivity: Option<f64>

The thermal conductivity of the substance in W/m.K

specific_heat_capacity: Option<f64>

The specific heat capacity of the substance in J/kg.K

density: Option<f64>

The density of the substance in kg/m3

Implementations

impl Normal

pub fn from_bytes(
line: usize,
bytes: &[u8],
model: &SimpleModel
) -> Result<Self, String>

Creates a new Normal by scanning bytes (i.e., a &[u8]) acquired from a text file

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

impl Normal

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new(name: String) -> Self

Creates a new Normal

All the required fields are asked by the constructor. The Optional fields default to None

pub fn index(&self) -> Result<&usize, String>

Gets the index field. Returns a Result because this field is optional and thus it might not be there.

pub fn thermal_conductivity(&self) -> Result<&f64, String>

Gets the thermal_conductivity field. Returns a Result because this field is optional and thus it might not be there.

pub fn specific_heat_capacity(&self) -> Result<&f64, String>

Gets the specific_heat_capacity field. Returns a Result because this field is optional and thus it might not be there.

pub fn density(&self) -> Result<&f64, String>

Gets the density field. Returns a Result because this field is optional and thus it might not be there.

pub fn set_index(&mut self, v: usize) -> &mut Self

Sets the index field.

pub fn set_thermal_conductivity(&mut self, v: f64) -> &mut Self

Sets the thermal_conductivity field.

pub fn set_specific_heat_capacity(&mut self, v: f64) -> &mut Self

Sets the specific_heat_capacity field.

pub fn set_density(&mut self, v: f64) -> &mut Self

Sets the density field.

impl Normal

pub fn thermal_diffusivity(&self) -> Result<f64, String>

Calculates the thermal diffusivity of the Normal

pub fn wrap(self) -> Substance