Crate model

This crate contains the data structure utilized for describing a SIMPLE building.

This structure and its API are built on top of an architecture developed during Germán Molina’s PhD dissertation. The purpose of this architecture is to keep the different physical domains of a building—e.g., thermal, visual, air flows, moisture—separated in the code but integrated at runtime. This allows for multi-physics building simulation and a couple of more interesting things while keeping the code maintainable and scalable.

Introduction

It contains all the supported elements of a building and methods for manipulating them.

This library includes a macro for generating automatic user documentation and a highly consistent RUST API. This macro also enables a scanner to produce functions for parsing a model from text files.

Similarly, it includes a rhai_api—i.e., the Control API—that allows manipulating certain elements of the building through portable scripts, allowing users to emulate automatic control or occupant behaviour.

Architecture of SIMPLE and the model and Simulation Engine

The basics of the architecture is that the building simulation is performed through different simulation modules, coresponding to the different domains to simulate (e.g., thermal, visual, etc.). However—because these domains tend to use different abstractions, methods, math and so on—integrating them in a maintainable way can be tricky. The architecture presented here attempts to mitigate this issue by keeping the domains separated in the code but tightly communicated at runtime.

Communication is achieved through the Model structure and its associated simulation_state. Specifically, after its creation, the Model and the simulation modules themselves all remains immutable, and all the changes are applied over the simulation_state. This implies that people developing different modules need to only understand a single API: This one. Also, they can retrieve information produced by other domains of the simulation by simply asking for it.

Example

use model::{Model, Space, SimulationStateHeader, SimulationStateElement};

/* INITIALIZE MODEL AND STATE */
// The model is simple
let mut model = Model::default();
// Internally, the State has a header (the element to which the data in it belongs)
// and its value. We will decouple them later.
let mut header = SimulationStateHeader::new();

/* POPULATE THE MODEL */
// Let's create a bedroom.
let mut bedroom = Space::new("my_bedroom");

// Let's put the bedroom in a model. This adds an index of the bedroom object, which
// we will use later
// (This also wraps the bedroom on an `Rc` and returns a clone of it)
let bedroom = model.add_space(bedroom);

/*  MAP THE STATE OF THE ROOM
   (whis will be done by SimulationModules, as they know what is needed; e.g., temperature or air speed)
*/
// We need to submit a SimulationStateElement associated to the SpaceDryBulbTemperature.
// We choose an initial value to 22.0 C
// Note that the space is identified through its index.
let space_index = 0; // We know this... it was the first to be added.
let state_index = header.push(SimulationStateElement::SpaceDryBulbTemperature(space_index), 22.).unwrap();
// Then we need to tell the space which index in the SimulationState contains its DryBulb temperature.
bedroom.set_dry_bulb_temperature_index(state_index);

// After building the model, we decouple the header of the simulation state
// (i.e., long and fancy names) from its data. The data is just a compact,
// memory-efficient and cool Vector of floating point numbers that will be passed
// around simulation modules. This structure is cheaper to read, write and move around.
let mut sim_state = header.take_values().unwrap();

/* EMULATE A FANCY THERMAL CALCULATION */

// Let's assume that—through a fancy calculation—we concluded that
// the current temperature is 24.0
bedroom.set_dry_bulb_temperature(&mut sim_state, 24.);

/* RETREIVE THE DATA FROM A DIFFERENT MODULE */
// Let's assume we need the dry bulb temperature in order to calculate the number
// of flies that might be in this room, bothering occupants... As biologists, we know
// nothing about room heat transfer. Fortunately, this has been calculated already
assert!( (24.0 - bedroom.dry_bulb_temperature(&sim_state).unwrap() ).abs() < 1e-5 );

Re-exports

pub use substance::Substance;

pub use hvac::HVAC;

pub use simulation_options::SolarOptions;

pub use error_msgs::print_error;

pub use error_msgs::print_warning;

Modules

boundary 🔒

Represents the boundary of a Surface (e.g. it can lead to the ground, outdoors or a space)

building 🔒

A Building Object that can conain spaces.

construction 🔒

A construction; i.e., a set of materials ordered from Front to Back

error_msgs

For printing warning and error messages to the user

fenestration 🔒

A surface that can potentially be opened and closed.

hvac

A collection of elements heating and cooling systems

infiltration 🔒

An infiltration rate for a Space

luminaire 🔒

A Luminaire

material 🔒

A material; i.e., a Substance with a certain thickness

model 🔒

The model itself

rhai_api

The module in charge of registering the control API.

scanner

The module containing the functions that allow parsing a Model from text files

simulation_options

For setting options in simulations

simulation_state 🔒

Contains the structure that has all the data that changes throughout the simulation

simulation_state_element 🔒

Contains all the possible elements that can be registered in the simulation state (e.g., Space Dry Bulb Temperature, Fenestration Solar Irradiance, etc.)

site_details 🔒

Some details of the site in which the building(s) is located

space 🔒

Represents a space within a building. This will often be a room, but it might also be half a room

substance

A physical substance with physical—i.e., optical, thermal—properties.

surface 🔒

A fixed (i.e., not movable) surface in the building (or surroundings). This can be of any Construction, transparent or not.

Structs

Building

This object is utilized to group Space objects together for metering and/or shared values. For example, the number of storeys and the ShelterClass will help defining the Infiltrations

Construction

An object representing an array of Materials, ordered from front to back

Fenestration

A surface that can potentially be opened and closed. It can be of any Construction and it does not need to be a hole in another surface.

Luminaire

A Luminaire

Material

The representation of a physical layer-Material. That is to say, a layer of a certain thickness made of a certain Substance

Model

A structure describing a set of built-environment objects.

SimulationStateHeader

The SimulationState is a Vector of SimulationStateElement objects. It is intended to be a quick-to-clone structure.

SiteDetails

Some information about the site in which the building(s) are located

Space

Represents a space with homogeneous temperature within a building. It is often actual room enclosed by walls, but it can also be more than one room. In this latter case, there will be walls within the Space, meaning that there are walls whose Front and Back boundary is this space.

Surface

A fixed (i.e., not movable) surface in the building (or surroundings). This can be of any Construction, transparent or not.

Enums

Boundary

Represents the boundary of a Surface

FenestrationPosition

Defines whether the Fenestration is fixed or openable.

FenestrationType

Defines whether the fenestration is a Door or a Window.

Infiltration

Infiltration is the unintended air exchange between an interior space and the exterior. In SIMPLE, infiltrations are added to the intended ventilation.

Output

Possible outputs to request from the simulation

ShelterClass

Indicates how sheltered is the building. This affects its infiltration

SimulationStateElement

The idea is to have a cheap-to-clone (or copy?) structure

SurfaceType

The kind of surface, useful for pre- and post-processing.

TerrainClass

This class modifies the wind speed in the site

Type Definitions

Float 🔒

SimulationState

The structure containing the data that changes during the simulation