Struct model::fenestration::Fenestration

pub struct Fenestration {

    pub name: String,
    pub vertices: Polygon3D,
    pub construction: String,
    pub category: Option<FenestrationType>,
    operation: Option<FenestrationPosition>,
    front_boundary: Option<Boundary>,
    back_boundary: Option<Boundary>,
    parent_surface: Option<String>,
    first_node_temperature: RefCell<Option<usize>>,
    last_node_temperature: RefCell<Option<usize>>,
    open_fraction: RefCell<Option<usize>>,
    front_convection_coefficient: RefCell<Option<usize>>,
    back_convection_coefficient: RefCell<Option<usize>>,
    front_convective_heat_flow: RefCell<Option<usize>>,
    back_convective_heat_flow: RefCell<Option<usize>>,
    front_incident_solar_irradiance: RefCell<Option<usize>>,
    back_incident_solar_irradiance: RefCell<Option<usize>>,
    front_ir_irradiance: RefCell<Option<usize>>,
    back_ir_irradiance: RefCell<Option<usize>>,

}

A surface that can potentially be opened and closed. It can be of any Construction and it does not need to be a hole in another surface.

Examples

.spl

{{#include ../../../tests/scanner/fenestration.spl}}

.json

{{#include ../../../tests/scanner/fenestration.json}}

Fields
name: String

The name of the sub surface

vertices: Polygon3D

An array of Numbers representing the vertices of the surface. The length of this array must be divisible by 3.

construction: String

The name of the Construction object in the constructions property of the Model object

category: Option<FenestrationType>

Defines whether a Fenestration is a Window, Door, or other. If none is given, the assumed behaviour is that it is a Window.

operation: Option<FenestrationPosition>

The opportunity for operating the Fenestration. If none is given, the window is assumed to be Fixed at Closed position.

front_boundary: Option<Boundary>

The front Boundary. No boundary means it leads to the exterior

back_boundary: Option<Boundary>

The back Boundary. No boundary means it leads to the exterior

parent_surface: Option<String>

The name of the surface containing this Fenestration, if any. A hole will be made in the parent surface in order to accomodate

Note that a Fenestration can be self contained (i.e., it can have no parent surface), which allows representing situation where the fenestration is very large and therefore there would be no area for wall.

Note: This field will not be serialized. This is beacuse there is no quarantee that the surfaces will be there before the fenestrations and thus there would be errors when deserializing. In any case, it is assumed that JSON models are read/written by machines—e.g., either by serializing a Model or by another kind of machine—and therefore the convenience of not having to write down the vertices around holes is not much needed.

first_node_temperature: RefCell<Option<usize>>last_node_temperature: RefCell<Option<usize>>open_fraction: RefCell<Option<usize>>

Index of the SimulationStateElement representing the fraction open in the SimulationState

front_convection_coefficient: RefCell<Option<usize>>back_convection_coefficient: RefCell<Option<usize>>front_convective_heat_flow: RefCell<Option<usize>>back_convective_heat_flow: RefCell<Option<usize>>front_incident_solar_irradiance: RefCell<Option<usize>>back_incident_solar_irradiance: RefCell<Option<usize>>front_ir_irradiance: RefCell<Option<usize>>back_ir_irradiance: RefCell<Option<usize>>
Implementations

impl Fenestration

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

Prints the docs for this object

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new<S: Into<String>>(name: S, vertices: Polygon3D, construction: S) -> Self

Creates a new Fenestration

All the required fields are asked by the constructor. The Optional fields default to None, and the Vec fields default to emptt.

pub fn category(&self) -> Result<&FenestrationType, String>

Gets the category field. Returns a Result because this field is optional and thus it might not be there.

pub fn operation(&self) -> Result<&FenestrationPosition, String>

Gets the operation field. Returns a Result because this field is optional and thus it might not be there.

pub fn front_boundary(&self) -> Result<&Boundary, String>

Gets the front_boundary field. Returns a Result because this field is optional and thus it might not be there.

pub fn back_boundary(&self) -> Result<&Boundary, String>

Gets the back_boundary field. Returns a Result because this field is optional and thus it might not be there.

pub fn parent_surface(&self) -> Result<&String, String>

Gets the parent_surface field. Returns a Result because this field is optional and thus it might not be there.

pub fn first_node_temperature_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the first_node_temperature within the [SimulationState].

pub fn first_node_temperature(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the first_node_temperature within the [SimulationState].

pub fn last_node_temperature_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the last_node_temperature within the [SimulationState].

pub fn last_node_temperature(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the last_node_temperature within the [SimulationState].

pub fn open_fraction_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the open_fraction within the [SimulationState].

pub fn open_fraction(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the open_fraction within the [SimulationState].

pub fn front_convection_coefficient_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the front_convection_coefficient within the [SimulationState].

pub fn front_convection_coefficient(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the front_convection_coefficient within the [SimulationState].

pub fn back_convection_coefficient_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the back_convection_coefficient within the [SimulationState].

pub fn back_convection_coefficient(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the back_convection_coefficient within the [SimulationState].

pub fn front_convective_heat_flow_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the front_convective_heat_flow within the [SimulationState].

pub fn front_convective_heat_flow(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the front_convective_heat_flow within the [SimulationState].

pub fn back_convective_heat_flow_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the back_convective_heat_flow within the [SimulationState].

pub fn back_convective_heat_flow(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the back_convective_heat_flow within the [SimulationState].

pub fn front_incident_solar_irradiance_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the front_incident_solar_irradiance within the [SimulationState].

pub fn front_incident_solar_irradiance(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the front_incident_solar_irradiance within the [SimulationState].

pub fn back_incident_solar_irradiance_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the back_incident_solar_irradiance within the [SimulationState].

pub fn back_incident_solar_irradiance(
&self,
state: &SimulationState
) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the back_incident_solar_irradiance within the [SimulationState].

pub fn front_ir_irradiance_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the front_ir_irradiance within the [SimulationState].

pub fn front_ir_irradiance(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the front_ir_irradiance within the [SimulationState].

pub fn back_ir_irradiance_index(&self) -> Option<usize>

Gets the index of the [SimulationStateElement] representing the back_ir_irradiance within the [SimulationState].

pub fn back_ir_irradiance(&self, state: &SimulationState) -> Option<f64>

Gets the value of the [SimulationStateElement] representing the back_ir_irradiance within the [SimulationState].

pub fn set_category(&mut self, v: FenestrationType) -> &mut Self

Sets the category field.

pub fn set_operation(&mut self, v: FenestrationPosition) -> &mut Self

Sets the operation field.

pub fn set_front_boundary(&mut self, v: Boundary) -> &mut Self

Sets the front_boundary field.

pub fn set_back_boundary(&mut self, v: Boundary) -> &mut Self

Sets the back_boundary field.

pub fn set_parent_surface<S: Into<String>>(&mut self, v: S) -> &mut Self

Sets the parent_surface field.

pub fn set_first_node_temperature_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the first_node_temperature within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_first_node_temperature(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the first_node_temperature within the [SimulationState] .

pub fn set_last_node_temperature_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the last_node_temperature within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_last_node_temperature(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the last_node_temperature within the [SimulationState] .

pub fn set_open_fraction_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the open_fraction within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_open_fraction(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the open_fraction within the [SimulationState] .

pub fn set_front_convection_coefficient_index(
&self,
i: usize
) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the front_convection_coefficient within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_front_convection_coefficient(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the front_convection_coefficient within the [SimulationState] .

pub fn set_back_convection_coefficient_index(
&self,
i: usize
) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the back_convection_coefficient within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_back_convection_coefficient(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the back_convection_coefficient within the [SimulationState] .

pub fn set_front_convective_heat_flow_index(
&self,
i: usize
) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the front_convective_heat_flow within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_front_convective_heat_flow(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the front_convective_heat_flow within the [SimulationState] .

pub fn set_back_convective_heat_flow_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the back_convective_heat_flow within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_back_convective_heat_flow(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the back_convective_heat_flow within the [SimulationState] .

pub fn set_front_incident_solar_irradiance_index(
&self,
i: usize
) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the front_incident_solar_irradiance within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_front_incident_solar_irradiance(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the front_incident_solar_irradiance within the [SimulationState] .

pub fn set_back_incident_solar_irradiance_index(
&self,
i: usize
) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the back_incident_solar_irradiance within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_back_incident_solar_irradiance(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the back_incident_solar_irradiance within the [SimulationState] .

pub fn set_front_ir_irradiance_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the front_ir_irradiance within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_front_ir_irradiance(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the front_ir_irradiance within the [SimulationState] .

pub fn set_back_ir_irradiance_index(&self, i: usize) -> Result<(), String>

Sets the index of the [SimulationStateElement] representing the back_ir_irradiance within the [SimulationState]. Returns an error if the object already has an index assigned for this field.

pub fn set_back_ir_irradiance(
&self,
state: &mut SimulationState,
v: f64
) -> Result<(), String>

Changes the value of the [SimulationStateElement] associated with the back_ir_irradiance within the [SimulationState] .

impl Fenestration

pub fn register_api(
engine: &mut Engine,
model: &Rc<Model>,
state: &Rc<RefCell<SimulationState>>,
research_mode: bool
)

Registers the Rhai API for the Fenestration

pub fn print_api_doc(dir: &str, summary: &mut str) -> Result<()>

Prints the Rhai API for the Fenestration

impl Fenestration

pub fn clone_loop(&self) -> Loop3D

Clones the outer [Loop3D] of the Fenestration

pub fn area(&self) -> f64

Gets the area, based on the [Polygon3D] that represents this Fenestration

pub fn is_operable(&self) -> bool

Can the fenestration be operated?