Struct model::substance::normal::Normal

pub struct Normal {
    pub name: String,
    thermal_conductivity: Option<f64>,
    specific_heat_capacity: Option<f64>,
    density: Option<f64>,
    front_solar_absorbtance: Option<f64>,
    back_solar_absorbtance: Option<f64>,
    solar_transmittance: Option<f64>,
    front_visible_reflectance: Option<f64>,
    back_visible_reflectance: Option<f64>,
    visible_transmissivity: Option<f64>,
    front_thermal_absorbtance: Option<f64>,
    back_thermal_absorbtance: Option<f64>,
}

Represents a physical material with common physical properties (e.g., timber, concrete, brick, glass). In other words, it cannot change its thermal/optical properties based on its internal state (e.g., it cannot change its specific heat based on temperature, like Gas or Phase Change Materials).

Examples

.spl

{{#include ../../../tests/scanner/substance_normal.spl}}

.json

{{#include ../../../tests/scanner/substance_normal.json}}

Fields
name: String

The name of the Substance. Should be unique for each Substance in the Model object

thermal_conductivity: Option<f64>

The thermal conductivity of the substance in W/m.K

specific_heat_capacity: Option<f64>

The specific heat capacity of the substance in J/kg.K

density: Option<f64>

The density of the substance in kg/m3

front_solar_absorbtance: Option<f64>

Solar absorbtance (from 0 to 1) at the front side (Front being the side closer to the first material in a construction)

Absorbtance is used instead of reflectance (which is used in visible radiation properties) to maintain coherence with EnergyPlus

back_solar_absorbtance: Option<f64>

Solar absorbtance (from 0 to 1) at the front side (Back being the side closer to the last material in a construction)

Absorbtance is used instead of reflectance (which is used in visible radiation properties) to maintain coherence with EnergyPlus… because in Thermal calculation we mostly care about how much is absorbed; in lighting we care mainly about how much is reflected.

solar_transmittance: Option<f64>

The front solar transmittance at normal incidence (from 0 to 1)

Please note that, contrary to all other properties, this property does depend on the thickness of the substance. So, in order to build a coherent Glazing, you’ll need to match this Substance with an appropriate Material

Transmittance is used instead of transmissivity (which is used in visible radiation properties) to maintain coherence with EnergyPlus… because in Thermal calculation we mostly care about how much is absorbed; in lighting we care mainly about how much is reflected.

front_visible_reflectance: Option<f64>

Solar absorbtance (from 0 to 1) at the front side (Front being the side closer to the first material in a construction)

Reflectance is used instead of Absorbtance (which is used in solar radiation properties) to maintain coherence with Radiance… because in Lighting we really care about how much is reflected; in thermal we care mainly about how much is absorbed.

back_visible_reflectance: Option<f64>

Solar absorbtance (from 0 to 1) at the front side (Back being the side closer to the last material in a construction)

Reflectance is used instead of Absorbtance (which is used in solar radiation properties) to maintain coherence with Radiance… because in Lighting we really care about how much is reflected; in thermal we care mainly about how much is absorbed.

visible_transmissivity: Option<f64>

The front solar transmittance at normal incidence (from 0 to 1)

Please note that, contrary to all other properties, this property does depend on the thickness of the substance. So, in order to build a coherent Glazing, you’ll need to match this Substance with an appropriate Material

Transmissivity is used instead of Transmittance (which is used in solar radiation properties) to maintain coherence with Radiance

front_thermal_absorbtance: Option<f64>

Front thermal absorbtance (i.e., emissitivy; from 0 to 1) (Front being the side closer to the first material in a construction)

back_thermal_absorbtance: Option<f64>

Back thermal absorbtance (i.e., emissitivy; from 0 to 1) (Back being the side closer to the last material in a construction)

Implementations

impl Normal

pub fn print_doc(dir: &str, summary: &mut String) -> Result<()>

Prints the docs for this object

pub fn object_type(&self) -> &str

Retrieves the type of object as a &'static str.

This method is useful for debuging models that contain multiple objects

pub fn name(&self) -> &String

Borrows the name

pub fn new<S: Into<String>>(name: S) -> Self

Creates a new Normal

All the required fields are asked by the constructor. The Optional fields default to None, and the Vec fields default to emptt.

pub fn thermal_conductivity_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of thermal_conductivity field or a default value given if the field is None. If no default is logical, use self.thermal_conductivity() instead, which returns an error

pub fn thermal_conductivity(&self) -> Result<&f64, String>

Gets the thermal_conductivity field. Returns a Result because this field is optional and thus it might not be there.

pub fn specific_heat_capacity_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of specific_heat_capacity field or a default value given if the field is None. If no default is logical, use self.specific_heat_capacity() instead, which returns an error

pub fn specific_heat_capacity(&self) -> Result<&f64, String>

Gets the specific_heat_capacity field. Returns a Result because this field is optional and thus it might not be there.

pub fn density_or<T: Display>(&self, module_name: T, default: f64) -> f64

Gets the content of density field or a default value given if the field is None. If no default is logical, use self.density() instead, which returns an error

pub fn density(&self) -> Result<&f64, String>

Gets the density field. Returns a Result because this field is optional and thus it might not be there.

pub fn front_solar_absorbtance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of front_solar_absorbtance field or a default value given if the field is None. If no default is logical, use self.front_solar_absorbtance() instead, which returns an error

pub fn front_solar_absorbtance(&self) -> Result<&f64, String>

Gets the front_solar_absorbtance field. Returns a Result because this field is optional and thus it might not be there.

pub fn back_solar_absorbtance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of back_solar_absorbtance field or a default value given if the field is None. If no default is logical, use self.back_solar_absorbtance() instead, which returns an error

pub fn back_solar_absorbtance(&self) -> Result<&f64, String>

Gets the back_solar_absorbtance field. Returns a Result because this field is optional and thus it might not be there.

pub fn solar_transmittance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of solar_transmittance field or a default value given if the field is None. If no default is logical, use self.solar_transmittance() instead, which returns an error

pub fn solar_transmittance(&self) -> Result<&f64, String>

Gets the solar_transmittance field. Returns a Result because this field is optional and thus it might not be there.

pub fn front_visible_reflectance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of front_visible_reflectance field or a default value given if the field is None. If no default is logical, use self.front_visible_reflectance() instead, which returns an error

pub fn front_visible_reflectance(&self) -> Result<&f64, String>

Gets the front_visible_reflectance field. Returns a Result because this field is optional and thus it might not be there.

pub fn back_visible_reflectance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of back_visible_reflectance field or a default value given if the field is None. If no default is logical, use self.back_visible_reflectance() instead, which returns an error

pub fn back_visible_reflectance(&self) -> Result<&f64, String>

Gets the back_visible_reflectance field. Returns a Result because this field is optional and thus it might not be there.

pub fn visible_transmissivity_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of visible_transmissivity field or a default value given if the field is None. If no default is logical, use self.visible_transmissivity() instead, which returns an error

pub fn visible_transmissivity(&self) -> Result<&f64, String>

Gets the visible_transmissivity field. Returns a Result because this field is optional and thus it might not be there.

pub fn front_thermal_absorbtance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of front_thermal_absorbtance field or a default value given if the field is None. If no default is logical, use self.front_thermal_absorbtance() instead, which returns an error

pub fn front_thermal_absorbtance(&self) -> Result<&f64, String>

Gets the front_thermal_absorbtance field. Returns a Result because this field is optional and thus it might not be there.

pub fn back_thermal_absorbtance_or<T: Display>(
&self,
module_name: T,
default: f64
) -> f64

Gets the content of back_thermal_absorbtance field or a default value given if the field is None. If no default is logical, use self.back_thermal_absorbtance() instead, which returns an error

pub fn back_thermal_absorbtance(&self) -> Result<&f64, String>

Gets the back_thermal_absorbtance field. Returns a Result because this field is optional and thus it might not be there.

pub fn set_thermal_conductivity(&mut self, v: f64) -> &mut Self

Sets the thermal_conductivity field.

pub fn set_specific_heat_capacity(&mut self, v: f64) -> &mut Self

Sets the specific_heat_capacity field.

pub fn set_density(&mut self, v: f64) -> &mut Self

Sets the density field.

pub fn set_front_solar_absorbtance(&mut self, v: f64) -> &mut Self

Sets the front_solar_absorbtance field.

pub fn set_back_solar_absorbtance(&mut self, v: f64) -> &mut Self

Sets the back_solar_absorbtance field.

pub fn set_solar_transmittance(&mut self, v: f64) -> &mut Self

Sets the solar_transmittance field.

pub fn set_front_visible_reflectance(&mut self, v: f64) -> &mut Self

Sets the front_visible_reflectance field.

pub fn set_back_visible_reflectance(&mut self, v: f64) -> &mut Self

Sets the back_visible_reflectance field.

pub fn set_visible_transmissivity(&mut self, v: f64) -> &mut Self

Sets the visible_transmissivity field.

pub fn set_front_thermal_absorbtance(&mut self, v: f64) -> &mut Self

Sets the front_thermal_absorbtance field.

pub fn set_back_thermal_absorbtance(&mut self, v: f64) -> &mut Self

Sets the back_thermal_absorbtance field.

impl Normal

pub fn thermal_diffusivity(&self) -> Result<f64, String>

Calculates the thermal diffusivity of the Normal

pub fn wrap(self) -> Substance

Wraps a Normal in a Substance enum